# internal/adapters/http/dto/response/user_response.go
package response

import "time"

// UserResponse represents the response with user data
type UserResponse struct {
	ID        string    `json:"id"`
	Email     string    `json:"email"`
	Name      string    `json:"name"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}